* commit 'accb0ab2bff5b61f1f385288a4d22f436b934c53':
  remove unnecessary test
  change type of lenr to u16
  change type of lenc to u16
  add package-lock.json
  update npm dependencies
  adjust debug output
  handle parse error of response data
  change log level
  adjust range test

+++ src/lib.rs
@@ -32,9 +32,9 @@ pub extern "system" fn CT_init(ctn: u16, pn: u16) -> i8 {
 pub extern "system" fn CT_data(ctn: u16,
                                dad: *mut u8,
                                sad: *mut u8,
-                               lenc: usize,
+                               lenc: u16,
                                command: *const u8,
-                               lenr: *mut usize,
+                               lenr: *mut u16,
                                response: *mut u8)
                                -> i8 {
     logging::init();
